# specs/001-specify-project-title/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Dynamic CrewAI Orchestration Platform API
  version: 1.0.0
  description: |
    Multi-tenant platform for creating, managing, and executing AI crew workflows.

    **Authentication**: All endpoints (except /auth/register and /auth/login) require JWT Bearer token.

    **Multi-Tenancy**: Tenant context is extracted from JWT claims. All data is tenant-isolated.

    **Rate Limiting**: Per-tenant rate limits enforced. See tenant quotas for details.
  contact:
    name: API Support
    email: api@crewai-platform.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.crewai-platform.com/api/v1
    description: Production
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Flow:
      type: object
      required: [name, nodes, edges]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/FlowNode'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/FlowEdge'
        input_schema:
          type: object
          description: JSON Schema for flow inputs
        output_schema:
          type: object
          description: JSON Schema for flow outputs
        version:
          type: integer
        status:
          type: string
          enum: [draft, published, archived]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FlowNode:
      type: object
      required: [id, type, position, data]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [agent, tool, llm, condition, input, output]
        position:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
        data:
          type: object
          description: Type-specific node configuration
    FlowEdge:
      type: object
      required: [id, source, target]
      properties:
        id:
          type: string
          format: uuid
        source:
          type: string
          format: uuid
        target:
          type: string
          format: uuid
        sourceHandle:
          type: string
        targetHandle:
          type: string
...
    ChatSession:
      type: object
      properties:
        id:
          type: string
          format: uuid
        crew_id:
          type: string
          format: uuid
        attached_flow_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
...
    PaginationInfo:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
paths:
...
  /flows:
    get:
      tags: [Flows]
      summary: List flows for current tenant
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, archived]
      responses:
        '200':
          description: List of flows
          content:
            application/json:
              schema:
                type: object
                properties:
                  flows:
                    type: array
                    items:
                      $ref: '#/components/schemas/Flow'
                  pagination:
                    $ref: '#/components/schemas/PaginationInfo'

    post:
      tags: [Flows]
      summary: Create new flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Flow'
      responses:
        '201':
          description: Flow created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flow'
        '400':
          description: Invalid flow definition
  /flows/{flow_id}:
    parameters:
      - name: flow_id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Flows]
      summary: Get flow details
      responses:
        '200':
          description: Flow details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flow'
        '404':
          description: Flow not found

    put:
      tags: [Flows]
      summary: Update flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Flow'
      responses:
        '200':
          description: Flow updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flow'

    delete:
      tags: [Flows]
      summary: Delete flow (soft delete)
      responses:
        '204':
          description: Flow deleted
  /flows/{flow_id}/execute:
    post:
      tags: [Flows]
      summary: Execute flow
      parameters:
        - name: flow_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                inputs:
                  type: object
                config:
                  type: object
      responses:
        '202':
          description: Execution started
          content:
            application/json:
              schema:
                type: object
                properties:
                  execution_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [queued]

  # Execution endpoints
...
  /chat/sessions:
    get:
      tags: [Chat]
      summary: List chat sessions
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChatSession'

    post:
      tags: [Chat]
      summary: Create chat session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                crew_id:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Session created
  /chat/sessions/{session_id}/messages:
    post:
      tags: [Chat]
      summary: Send message to crew
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                attached_flow_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Message sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message_id:
                    type: string

  # LLM Provider endpoints
...